Fix release query to get latest releases
Let's only show the last 50 releases as we really don't need to show 100
of them, with many of them being out of service/very old. Also, adding
the page number seems to have fixed the problem that it wasn't showing
the very latest releases? Or maybe it was a cache issue? We'll know next
time we update a release :)

Signed-off-by: Phil Estes <estesp@amazon.com>